chore(firebase_analytics): update imports to match existing plugins
