+++ packages/firebase_analytics/firebase_analytics/example/lib/main.dart
@@ -8,7 +8,6 @@ import 'dart:async';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:firebase_analytics/observer.dart';
 
 import 'tabs_page.dart';
