+++ packages/firebase_analytics/firebase_analytics/lib/src/observer.dart
@@ -3,10 +3,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:flutter/services.dart';
-import 'package:flutter/widgets.dart';
-
-import 'firebase_analytics.dart';
+part of firebase_analytics;
 
 /// Signature for a function that extracts a screen name from [RouteSettings].
 ///
