+++ packages/firebase_analytics/firebase_analytics/test/observer_test.dart
@@ -12,7 +12,6 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:flutter/widgets.dart';
 
 import 'package:firebase_analytics/firebase_analytics.dart';
-import 'package:firebase_analytics/observer.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
